Optimized ExeUnpacker.
The bit reading code was using heap allocation for every bit read, so it
was very slow. I replaced it with a stack-allocated array and a dynamic
index (bits used), and it runs much faster now. The timing in Visual
Studio's debug mode shows it went from about 4600ms to 205ms (in release
mode it's only 6ms). This makes the program startup much faster when
debugging is necessary.