Added Game::basePath member.
This path variable is obtained from SDL since it's platform-specific,
and is mostly meant for use on macOS, where the ".app" format is a
container for the executable and can manipulate the current working
directory based on certain user settings. Also put the local "icon"
variable into a Surface wrapper to avoid needing an explicit
SDL_FreeSurface().