Added support for relative and absolute ArenaPath.
I think it's correct... but anyway, the program checks if ArenaPath is
relative or absolute and prepends the base path if it's relative. I'm
also in the process of adding code that will use a local
"options/options.txt" if it exists (for development purposes), and use
the options.txt in the options pref path instead (creating it if it
doesn't exist).