Refactored voxel rendering.
The software renderer's behavior hasn't changed, but the voxel drawing
code now follows a more organized path of execution, and more data is
being reused as well. It will be much easier to implement occlusion
culling with this new layout. It looks like the annoying texture
artifacting problem was fixed when I revised the drawing range
calculations, so I guess the software renderer looks slightly better
than it did before.