Updated options.txt.
Increased cursor scale to be closer to what it looks like in the original game, and lowered sound volume by quite a bit since some of Arena's sounds (like fanfare) can be really loud. The cursor scale isn't dependent on window resolution, so the user can still tweak it to the desired value.