Added some shading to the renderer.
The renderer now applies ambient lighting and sunlight to textures. I
think deviating from Arena's look should generally be avoided at this
point, so I'm clamping the sum of the ambient and sun components to 1.0
to keep the colors "mild" like they were in the original. I'll probably
also do this with street lamps and spells because it looks like the sum
of light intensity in Arena never exceeds 1.0 for any pixels.