Added a comment in Debug::crash().
There is a special case for macOS .apps when the program encounters a
fatal error; they aren't listening to the console, so they never hear
any getchar(). Therefore, showing a messagebox is a simple fix to the
problem, but in the future, a logging mechanism should also be in use.