Refactoring cursor code.
This is an in-progress change for having the renderer decide how to draw
the current cursor instead of having every panel do it. The active panel
will be asked for its cursor during the game loop (sometimes simple,
sometimes a function of mouse position, etc.), and the renderer will
draw it. The cursor alignment is important for determining how to offset
the cursor's position when drawn (normally it is not offset any, when
using 'top left').