Done with cursor code refactoring pass.
All cursor drawing code now resides in Game::render(), and panels do not
have any cursor drawing code, only a get method for the current cursor.
I think this is a step in the right direction, but it could always be
better.  Next, I would like to investigate pop-up functionality.