Initial work on sub-panel functionality.
This should lay the groundwork for pop-up interfaces to be used all over
the program. A sub-panel exists on top of the main panel and takes
control of event handling and game ticks. Multiple sub-panels can be
stacked on top of each other, in which case the top-most one is the
active one.