Added some renderer methods.
These two methods transform rectangles between the native and original
coordinate systems. At some point in the future, I'd like to remove the
program's dependence on the so-called "original" frame buffer, since it
was just intended to be a convenience for quickly prototyping screens,
but it's not as necessary now.