Added Clock class.
This is a general-purpose 24-hour clock that will be used for keeping
track of the current time of day. I expect that the tick() method will
be used fairly frequently, and that the incoming delta time value most
likely won't be the exact frame time due to time scaling (i.e., each
game minute is one real-time second or something).