Added modulo to wall texel access.
There was a very slight chance that the vertical texture coordinate
would be 1.0, and it would access a texel outside of the texture. Using
modulo is a quick fix for now, but in the future, I think maybe using
std::nextafter to clamp the U and V coordinates themselves would be a
better solution.