More work on RichTextString.
It now stores pointers to each of its character surfaces. I wasn't sure
if SDL_Surfaces should be so tightly coupled with a rich text string
(since it's not actually the class doing the rendering), so I'll just
consider it an optimization (since the surface lists are used in
dimensions calculation as well as text box generation).