Fixed bug with province map closest location.
The closest distance calculation was experiencing an integer overflow
when the mouse was near the top-left corner of the screen, because the
default closest position value (-SHORT_MAX, -SHORT_MAX) was too large.
Reducing it to (-1000, -1000) seems to work fine. A better solution
might be to make the closest position be a nullable value.