Improved delta time calculation in game loop.
The calls to SDL_GetTicks() and SDL_Delay() have been replaced with
std::chrono functions, and the units of time have been changed from
milliseconds to microseconds. The delta time calculation is now precise
enough to measure fractions of a millisecond, and the displayed frames
per second is more accurate as well.