Added partial support for text triggers.
Text triggers are now activated when the player walks into a voxel. They
are just displayed in the console for now, but they will be displayed
on-screen soon. I'm not sure if I like the text trigger API I wrote for
the WorldData class, so maybe that will change sometime.