Added text trigger pop-ups to game interface.
The game now displays text triggers to the screen when the player enters
certain voxels. The current on-screen duration is just a guess (at least
2.5 seconds, or 1 second for every 20 characters). Some of these
triggers only display once, so the "on voxel enter" trigger is disabled
for them afterwards. The other text trigger data (keys and riddles) can
be loaded, but they are not in use yet.