Fixed some voxel-related bugs.
The default height of voxels is now 128 instead of 100, and texture
coordinates are now being divided by the new value as well (previously
it was causing out-of-range texture accesses in the renderer). 128 is
the official size of voxels in "Arena units" (i.e., there are 128 units
of distance per voxel).