More work on .INF code (floor-related).
The first group of floor/ceiling definitions is now being read in. I'm
not sure exactly how the box cap and ceiling textures are expected to be
laid out in memory, and there are also some "loose" texture filenames
that don't have any associated ID, so there are still some more things
to figure out.