Reorganized some .INF code.
All of the state structs have been moved to an anonymous namespace, and
their enum classes are contained within each. Instead of writing virtual
classes for the parser, I decided to write lambdas instead. This way, I
didn't need to worry about making the structs be friends of INFFile. I
also fixed a bug with the last *TEXT section that wasn't getting written
to the INFFile, so now the messages about "resting in niches" appear in
the start dungeon.