More work on .INF floor state.
Added some code in the .INF parser to catch the cases in floor parsing
where a "loose" texture name might be read without an associated *BOXCAP
section. Also added a +1 to the floor texture ID when creating the voxel
data structures for the voxel grid -- this seems to fix the floor
textures. I'm not sure how Arena stores its textures, but I have a
feeling the floors and walls are in separate arrays.