Added load methods for two more .MIF tags.
The FLAT and INNS tags are found in a couple .MIF files; the first in
the Imperial Palace (IMPPAL.MIF) and the second in the wilderness
(WILD.MIF). They might occur in other places as well. Their purpose is
currently unknown, but it looks like FLAT has references to sprites like
trees and things that would normally be in an .INF file.