Added another edge case to .INF parser.
There is at least one riddle in the .INF files that contains an empty
line (first TEXT in LABRNTH2.INF), so the parser needs to not skip those
lines. Also added a check for *TOP in the FLOORS section, although I
don't know what it's for. It only appears in LABRNTH1.INF and
LABRNTH2.INF.