Cleaning up some .INF code.
I changed several of the texture name members to be indices into the
textures vector instead, since they were just duplicates of names in
that vector. I also changed the return value of several of the get
methods from const std::string& to const int*, since the interface
should expose an index only if it exists, and null seems like a
reasonable value for representing "none".