Added door voxels.
The doors from .MIF files are now present in the world. They are
considered "voxels" in Arena's engine, even though they swing on a
hinge. I think it does some kind of texture coordinate interpolation or
something. I'm not sure if I want to implement it differently in my
renderer, so they appear as voxels for now.