Refactored options code.
The options.txt file has been replaced with an options-default.txt file
(similar to OpenMW). Any user-specific changes to settings now go in an
options-changes.txt file in the user's prefs folder (i.e.,
AppData/Roaming/OpenTESArena). It's not perfect yet because the changes
file should be created by a wizard program, not on the program's first
run (or when the file doesn't exist).