Changed game pointer to reference.
The game pointer should never be null, and it doesn't need to be
polymorphic, so there's no reason it can't be a reference instead. The
Game class now has its copy constructor and assignment operator deleted
to make sure only one instance of Game is alive at a time (it doesn't
need to be moved either, but deleting that is optional).