Changed unnecessary pre-increments to post-increments.
No need to explicitly use a pre-increment if the result of the
expression is not used. Post-increments are generally easier to read,
and they should be the default form of incrementing and decrementing
values (i.e., i++). Iterators are the only exception to this rule, and
pre-increment should be used by default for them to prevent unnecessary
copies (i.e., ++iter).