Fixed depth bug with flat rendering.
Flats now have the correct depth value for each pixel column. The
problem was caused by interpolating between a flat's left and right Z
values, which turned out to not be the right way to calculate depth at
some arbitrary percentage across the flat, and it resulted in artifacts
like the bottoms of flats going through the ground. There might
eventually be another refactoring of flat code at some point since it
appears that all flats face the camera, contrary to what I previously
thought, so they could probably be drawn as quads instead. Doors, store
signs, and fences are actually a part of voxels, not flats, as it turns
out.