Added some test city functionality.
The .INF filename for the quickstart test city is now a function of the
.MIF file, as well as climate, weather, and the world type. Also made
the music play based on location (although this is just an early test),
and made the fog distance a function of the weather. Also changed some
other things like start points and interior/exterior definitions,
although none of it is perfect yet.