Refactored some .MIF, .INF, and font code.
All .MIF files now have a fixed array of four starting points. Most of
these points will be (0, 0), but this ensures that no out-of-bounds
accesses occur, and that the right one is chosen for the level index.
Also replaced some out-of-range exceptions in INFFile::getBoxSide() and
font code with a warning message and default value. Also switched the
.INF parser's default mode to "Floors" instead of "None", since the last
staff piece dungeon is missing that tag even when it's needed.