Swapped grid X and Z dimensions.
OpenTESArena is using a different coordinate system than Arena to make
the 3D math a bit more intuitive/modern (i.e., +X is north instead of
west, and +Z is east instead of south), so a transform is necessary to
change between systems. The width (X) and height (Y) of a level in Arena
are the depth (Z) and width (X) in OpenTESArena, respectively.