Simplified panel switching code.
When changing the current panel or sub-panel, the caller no longer needs
to allocate a std::unique_ptr at the callsite. Instead, they just need
to give a template argument for the type of panel, and then pass the
other arguments normally. Behind the scenes, it's using a variadic
template, and it passes the arguments onto the std::unique_ptr
constructor via std::forward().