Added skeleton for occlusion culling implementation.
Occlusion culling helps solve a number of performance problems. For
example, it eliminates overdraw completely, removes the need for
depth-testing opaque drawing ranges, and occasionally lets the ray
casting loop return early. I'm not sure if it's *technically* occlusion
culling, but it serves the same purpose (not drawing things behind
opaque pixels while concurrently avoiding a depth test). Just the
OcclusionData::clipRange() and update() methods need to be implemented
in order for occlusion culling to work, and to test it visually, depth
testing should be disabled in each of the pixel-drawing methods. The
renderer will likely need another redesign in the future to allow the
drawing of transparent voxels from back to front, instead of everything
from front to back. Until then, depth-testing for opaque pixels will
still be necessary.