Small fix to edge voxel texture IDs.
I'm pretty sure that edge voxels use the same texture index format as
door voxels (the lowest 6 bits). This means that the orientation must be
two bits (which works just fine). There is one edge voxel in
IMPERIAL.MIF with a texture index of 0 (which becomes -1 or 255 after
the subtraction). Since Arena only has 64 texture slots, I'm assuming
that this is a silent bug in the original game, and I think clamping the
texture index positive is a good enough solution (the texture appears as
gravel instead of solid gray).