Improved voxel type data for solid voxels.
When generating walls on the main floor, the LevelData code now checks
the texture index against several indices in the .INF file to determine
the voxel type (solid, menu, level up/down, etc.). It's not perfect yet,
but it lets various blocks now have the proper color on the automap.