Added secondary render method to game loop.
The new renderSecondary() method is only called for the top-most panel.
The idea is that, instead of having each panel contain extra state for
determining when to draw certain things (pop-ups, text, etc.), the game
loop decides that for the panel, and the panel just has to have the
right code in the right method. This helps keep us from duplicating code
for things like pop-ups, and keeps panels from having a lot of if/else
statements. The design itself could certainly improve, but it works well
enough for now.