Flipped U texture coordinate for far face on edges.
This makes edge voxels like store signs look different on both sides
instead of the same, which seems to be the expected behavior. Although
there are some signs with words on them that won't look right from one
side (like TZTTAV.IMG), the general case will be correct (so that some
signs like DZTTAV.IMG won't appear to be floating in the air).