Added ranged weapon animations.
Fortunately I was able to reuse most of the WeaponAnimation class. Just
had to add some conditions for whether the animation was for a ranged
weapon. The sheathing and unsheathing states for a ranged weapon are
instantaneous, so they're skipped in setState() and the animation goes
straight to the subsequent state.