Limit ranged attack to be above game world interface.
The mouse must be above the game world interface in order for arrows to
fire. In the original game, it's limited even further to just within the
center "X" region, but personally, that seems a little inconvenient, so
expanding that area seems like a more desirable solution (with no real
effect on gameplay).