Replaced some std::memcpy()'s with std::copy.
std::copy is the more C++ way of copying bytes, and it should always be
as fast or faster than std::memcpy(). I'm leaving SDL_memcpy() calls
alone because they mostly deal with void pointers, and it would look
worse to change them to std::copy() with lots of type conversions.