Redesigned executable data loading.
The aExeStrings.txt file has been greatly simplified, and now only holds
offsets to the beginning of each global variable (instead of every
value). The ExeData class replaces ExeStrings and contains member
definitions for the data coming in from the executable. These changes
take a lot of the busy work out of the executable data definitions. The
concept was to give the program the ability to "view" some of the global
data as it was seen by the original application, and I think it's
working out much better now.