Replaced unique_ptr initializations with init().
Previously I used unique_ptrs to control the initialization order of
class members that needed constructor parameters, but now that the
design is fairly established, it seems using init() is the better way to
go in the long run. I wanted to avoid having initialization bugs due to
forgetfulness (so things would be null instead of garbage), but using
init() is the more fitting option now (especially since most of the
class members shouldn't be nullable to begin with).