Added rough draft callback support for level up voxels.
Not sure yet how I want to do this, but the general idea is to allow the
game world panel to call a custom function when entering a transition
voxel, so it can handle things like choosing a random location for the
player when leaving the start dungeon. By default it would simply
decrement the world's level index.