Added sound resampling option.
The user can now choose the desired sound resampling option from the
options menu (default, fastest, medium, or best). The setting itself is
stored as an integer, but the program interprets it differently
depending on the implementation. Also copied alext.h and efx.h into the
repository per @kcat's suggestion (so the files will be found on all dev
systems).