Removed PPM file support.
Since SDL already has BMP support, and the only thing PPM was used for
was the program icon, there's no need to have PPM support anymore. It
was kind of an odd idea to begin with, and BMP is more common anyway. I
don't think we need PNG support, though, since there's not really
anything that needs it.