Replaced some unnecessary at()'s with front() and back().
Some at() calls are guaranteed to be within bounds (i.e., because of
code right before them adding elements at the back), so front() and
back() can be used instead. In any case, I still want to keep using at()
everywhere, even if it's redundant, because it makes the code more
robust for hardly any cost (and it's easier to type than [], in my
opinion).