Added rough draft random dungeon generation.
There is now a test option for random dungeons on the main menu. It
generates levels with 64x32 dimensions for named dungeons and levels
with 64x64 dimensions for wilderness dungeons. No LEVELUP or LEVELDOWN
transition blocks or triggers yet, and I think the Location class needs
to be redesigned soon so it can hold a location ID (city ID or dungeon
ID).