Added case-insensitive open() to VFS.
There are some files with different casing between Arena's floppy and CD
versions, so a case-insensitive open method is required in order to open
them with the same filename on Unix systems. The new method isn't truly
case-insensitive -- it just fits Arena's problem space (which is either
all caps or the first character upper-case with the rest lower-case).