Removed unnecessary destructors.
Following the Rule of Zero, destructors only need to be user-defined if
they are explicitly freeing something. Since the vast majority of the
project uses standard library types for memory management, this makes it
almost completely unnecessary for me to write destructors (except in the
case of things like SDL and OpenAL Soft because they use C types and
integer handles).