Added travel pop-up text on province maps.
When clicking on a location, the travel text pop-up is displayed and
shows details about the selected location. Currently there seems to be a
bug in the global point calculation, since the travel distance is
correct but the travel days is not, and clicking on some locations
causes an error because their global point is not in any province's
global rectangle.