Added nullable travel data to WorldMapPanel.
To have the world map remember a province map's selected location, it
needs to store a nullable member for that data. This member is passed
between the WorldMapPanel and ProvinceMapPanel as needed. This change
also fixes a small issue in Button::click() that was causing the passed
arguments to be copied instead of forwarded.