Switch to world map when leaving main quest dungeon.
When the player enters a transition voxel that would put the level index
out of range, open the world map instead. This has the side effect of
making the shift gate in the starting dungeon open the world map when
testing, which doesn't really matter since it's a test world (the actual
starting dungeon operates normally because its custom function is set
through character creation).