Added city arrival pop-up text.
When the player arrives at a city, it shows a pop-up of their travel
details and a description of where they are now. City-states have unique
strings, while towns and villages choose randomly from a set of three
strings each. Still need to implement the ruler title and name
replacement.