Added search functionality to ProvinceMapPanel.
The search button on province maps now works, and displays a text entry
pop-up. Depending on the input, it will either attempt to select the
given location as the travel target, or it will display a list of
locations to choose from. I decided to make the locations list
alphabetized for reasons described at ProvinceSearchSubPanel.cpp(226).
Currently there's no scroll bar for the list and no blinking cursor for
text input, but those will be added later.