Read classes list UI layout into ExeData.
The classes list UI member determines how the the scroll buttons and
list box of class names are positioned in character creation. On a
different note, the aExeStrings.txt file is starting to have name
collisions now, so it might be a good idea to implement namespace
support in the KeyValueMap class.