Improved province map searching.
The search function now supports approximate matches, so the player can
type in things like "dagger" and it will find Daggerfall. I'm fairly
sure this deviates from the original behavior, but the original search
function is not very good in my opinion and it's something that could be
improved here.