Added camera pitch limit option.
This lets the player set how far above or below the horizon they want to
be able to look in modern mode. The max camera pitch limit must be less
than 90 degrees because the camera cannot look straight up or down in
the renderer (and also because it'd break the vector cross product).