Revised world projection and ceiling height definitions.
The camera's up vector is now multiplied by 1.20 to account for the 20%
difference in screen height due to tall pixels. The default ceiling
height is now 100 instead of 128 (this is the correct value for sure).
The player's height is now the same as the original game (60 / 128, or
0.46875). Revised some of the collision hack so it takes the ceiling
height into account when determining the player's Y voxel coordinate.