Improved weapon fitting in modern mode
I changed the offset of the weapons depending on which side of the screen they intersected with. This way I could preserve the original width without causing the weapons to cut off. There can be some improvements made to the offset values, I didn't measure them exactly to the pixel.