Redesigned world data.
The world data is now separated into interiors and exteriors. Each
subtype has unique properties (like interiors having a list of levels),
so it makes sense to have each kind in their own file. There may be more
redesigning in the future (i.e., separating interiors from dungeons and
cities from wilderness). Exteriors are now designed to store state for
when the player is in an interior at that location (so they can be
placed at the interior entrance correctly).