Use strings from executable for menu .MIF prefixes.
Menus that have a mapping to a .MIF filename prefix (like EQUIP or
NOBLE) will fetch it from the locations struct in the ExeData class. If
a menu doesn't have a mapping, then an empty string is returned and the
caller knows not to generate a full .MIF filename. The caller should
also modify the PALACE string in case it's a town or village palace
(TOWNPAL, VILPAL).