Treat transparent voxels as solid for ray casts.
When the player tries to right click a building entrance when a
transparent voxel (i.e., an archway or hedge) is in the way, the
transparent voxel will block the ray cast. If the player stands in the
same voxel as the transparent voxel, it will not block the ray. This
matches the original game's behavior.