Added palace entrance support.
Clicking on a palace gate now brings the player to the palace interior. The Imperial City palace is a special case because it brings the player to IMPPAL.MIF which is actually a dungeon. Last thing to do is fix the reversed horizontal texture coordinates (probably with a "flipped" boolean for type 0xA voxels).