Removed minimum resolution scale limit.
The options menu still limits the value between 0.10 and 1.0 but the value can be set anywhere between 0 and 1 in the options file provided it's greater than zero. The only practical application of this is if people running on a 2160p or higher screen want to have the render resolution under 10% of the fullscreen resolution (i.e., under 384x216).