Initial work on chunk definition.
Each chunk in Arena is 64x64 blocks with a variable height depending on whether it's an interior or exterior. The original game probably doesn't store blocks above Y=2 for exteriors since they're all just duplicates of Y=2, but in this engine, every voxel will be uniquely addressable for each 3D coordinate in the chunk.