Initial work on chunk sets.
Chunk sets hold all the currently active chunks, and chunks can be added and removed by a caller at any time (for instance, when the player crosses a chunk boundary, causing new chunks to be added and old chunks to be removed). The class offers iteration so that all active chunks can be queried.