Reworking render threads to act as background threads (unfinished).
Going in the direction of having all render threads always on in the background so they do not incur the cost of thread creation every frame. They will do thread synchronization with the main thread via std::condition_variable and receive notify signals for when to go to the next step of rendering. Currently these changes deadlock because of something with condition variable notifications and locks not behaving in the way I expect them to.