Added DebugAssertMsg() macro.
An alternative to DebugAssert so I can still have heavyweight asserts that run in release mode and so I can choose to have messages or not, since messages currently cause heap allocation (which isn't a big deal; it was in the interest of convenience anyway). DebugAssert() now just checks a boolean and doesn't take a message.