Progress on revising visible distant sky.
Trying to fix visible distant objects when parallax is on so the renderer can handle objects that are partially behind the camera. It should get the slice of the shape's angle range that is within the camera's angle range and use that for texture coordinates. It shouldn't break when part of the shape is outside the projection plane.