Textures are rendering (almost) correctly.
Uses cubemap array textures to store all the voxel textures. Ceilings are not in the correct orientation and raised sections are full voxels. Chasms are not right, but that will require an adjustment to the vertices for chasms only which could prove troublesome, but I have some ideas. Also need to get level transition hooked up to the map generator.