Added moon shader, revised visible distant objects layout.
The signal colors of moons now use the sky gradient color like the original game. Visible distant objects are now in a struct with one vector and lots of start/end indices pointing into the vector for each object type. This was necessary so that moons could have a different pixel rendering method called for them.