Let Chunk class manage voxel data.
I think it makes sense for each chunk to have its own set of voxel data definitions. This should make it so each voxel can be one byte instead of two because there are generally fewer unique voxels in a chunk than a voxel grid due to the smaller dimensions.