Reworked error handling of asset loading.
Instead of naively using assertions to stop the engine on asset-loading failure, a boolean is returned from init functions to specify whether a value was initialized properly, and the caller can handle it as they see fit. Even if an error is not handled, it will still show up in the console log.