Added a simple scrolling inventory list box.
The inventory list box scrolls as expected and is filled with test values. The coordinates of the list box and the space between items are approximations; they look ok but I'm sure they aren't pixel perfect to the original at the moment. The color is (I believe) the equipped item color, and I'm almost certain the font choice is correct.