More progress on wild interior name generation.
Wilderness interior names are almost all correct now. The remaining cases occur likely because I assumed that the name of a tavern/temple entrance depends on the chunk the voxel is in, not the chunk the player is in (I figured there wasn't a big enough difference to matter, but that seems to have been wrong). Not sure how to approach this with the current name generation method (just running the algorithm over the entire wilderness), so will probably need to rework some things in 0.11.0 as part of the planned chunk redesign.