Finished implementing fading voxels.
The addition of this feature brings out the simple nature of the renderer, since now the opaque voxel shaders are more expensive because they have the potential to be in a fading state. In my planned re-design of the renderer, I want to investigate shader permutations and if there's a way to not have an explosion of similar-looking pixel render loops.