Practice with SSE and AVX instructions.
Committing as a reference for later so I have an idea of how software renderer shaders could be vectorized. Not ready for use because it doesn't have the special case loop afterwards to fill in any gaps due to the large stride in the vector loop. AVX is commented out in case it's not supported by default on all compilers.