More work on directional animations.
Not satisfied with using an angleID for registering textures to a particular angle. Would rather have the texture-adding method just push texture lists into a list with order assumed rather than determined by an integer. The angle percentage would be used to index into that list at the correct spot.