Start using if constexpr for shader permutations.
It appears that template parameters and if constexpr can move the weight of maintaining efficient shader permutations to the compiler (hurray!). Ultimately I don't expect there to be more than a few dozen permutations of all shaders combined; the number will probably grow the most from things like light count on a surface (if that's even a good way to do it... kind of sounds like it. A unique shader for 1 light, 2 lights, etc.).