Fixed the rayDirection calculation, mostly, by having it take into account y-Shear rather than pitch from the camera. This is much closer to how the game actually does it, but it's not quite perfect.
Rewrote the raycast algorithm based on the commented-out 3D-DDA implementation in the software renderer. I also added further raycast logic for voxel types that aren't all there (specifically raised platforms and diagonal walls). These are working for the most part, though testing might be done to see if more work is needed for edges, chasms, and transparent walls.

I'll still need to add logic to the raycast, after checking for voxels, to see if the user is, in fact, clicking on an entity just in front of a voxel.