Added logic for selecting entities. This logic is applied after selecting voxels.
When selecting an entity, if you hold G then it will pop up a message telling you that you selected an entity. Entity selection has been made pixel-precise, but that may not stay permanently.

As part of the pixel-precise selection logic I added a function to the Renderer and SoftwareRenderer classes to get the color of a texture at a specific texture coordinate. This may be changed to a boolean function called isSelectable to simplify its purpose as it's rather specific to the purposes in this new selection logic anyway.