Added some optimizations to the Entity selection logic. It now builds a map of all entities to voxels and only ray tests the entities that cross the voxel that is being tested.
I also moved the common code shared between rayCast in Physics and updateVisibleEntities in the SoftwareRenderer to a single getEntityVisibilityData method in the EntityManager. This function fills out a structure that stores the entity pointer, as well as the animation state and position info of the entity that's needed for ray testing in the software renderer as well as in the raycast function of the physics class.

Removed an unnecessary clamp function from the MathUtils namespace, and replaced the reference to it in the DistanceBetweenLineSegments with a reference to std::clamp