Removed the debug cursor that I accidentally checked in.
Added an intersection calculation to the Diagonal Wall v rayTest

Fixed the code formatting for the function declarations and definittions to have the & on the name rather than the type. This will likely still need to be fixed for local variables.

Changed my functtions that take const double & to just take double. No need to pass by constant reference when passing by value is just fine.

Changed RayPlaneIntersection to use Constants::Epsilon rather than a magic number.

Added a static_cast<float> to anglePercent calculation to get rid of a double to float implicit cast warning.