I added screenPointToRay but it required making the SoftwareRenderer::Camera struct public so that it could be instantiated and passed in from the GameWorldPanel.
Recombined the modern and classic UI ray calculation and put that in the screenPointToRay method.

Changed the lambda to get the rayDirection in GameWorldPanel::HandleClickInWorld(...) to only calculate the parameters for the new SoftwareRenderer::screenPointToWorld function, then call that function to do the calculation.

I have verified that this change works in both Modern and Classic UI modes.