Added a large number of unit tests for the physics colliders. This allowed me to find and fix quite a few logical bugs. A great many more will be needed before I feel comfortable enough with the colliders to begin incorporating them into gameplay.
These will be removed sometime in the future, but hopefully not until all of the physics side of things is solid. Playtesting can help us verify that things are working as we want, but this is a much faster way to ensure that, in all edge cases, things are behaving as expected.