Initial implementation of water and lava rendering.
Dry, water, and lava chasms are now rendering. This first pass has the chasm texture stretched across the whole screen. It appears there is no perfect solution to showing the chasm texture because of its nature in the original game, but there could be graphics options in the future to tweak its appearance.