Added periodic creature sounds.
Each creature now plays their sound centered on themselves and waits a random handful of seconds before playing again. This is a quick and dirty implementation and is missing a couple things, like updating the sound position in AudioManager::update(), and having some kind of probability function based on proximity of similar entities.