Allow null data pointers in buffer views.
Since MSVC release builds seem to allow std::vector::data() to return nullptr for empty vectors, the buffer views cannot assert that their data is always non-null. It should be enough that the caller reads the buffer count and does not attempt to dereference anything if the count is 0. This change fixes a crash in the renderer when there are no lights in the view frustum in a release build.