Added voxel def functions to LevelDefinition and LevelInstance.
Played around with this idea a little bit. It seems to make sense to keep all of the LevelDefinition's voxel definitions constant and read-only, and any new ones needed _after_ level generation (like new chasm permutations caused by the player deleting a floor) would go in the LevelInstance. When the code wants to know what a voxel is, it asks the level instance. If the level instance doesn't have that voxel definition, it asks the level definition.