Replaced MusicName with MusicLibrary.
All instances of music in the game now reference the music library which loads music definitions from a file. This makes music a lot easier to mod since every type of music in the game (character creation, dungeons, taverns, etc.) can have any set of MIDI files assigned to it.