Redesigned texture manager interface.
The texture manager can now fail to look-up a texture. Also added native support for 8-bit images. The ideal usage pattern now involves passing a texture ID obtained previously for fast look-up. The concept of a built-in palette has been removed and the caller needs to get the texture's palette ID before trying to get that texture's ID. Need to apply all of these changes to everywhere in the engine that uses the texture manager.