Fixed several dangling reference bugs.
Need to get texture IDs in advance of any texture references that are obtained before a texture ID look-up and are used after. This isn't ideal but once I get around to a big Panel refactor with bool init() instead of constructors, it'll be a lot easier to initialize the texture IDs beforehand.