Added EntityRef wrapper.
Gives a safer option of accessing entities in the EntityManager than always working with raw entity references that might get invalidated when they change chunks. Also, passing an entity's type in the entity getter functions helps speed up the look-up a bit because it knows which static/dynamic entity groups to search.